Move Unit Entity to the Unit/Entity Folder
- Refactoring based on ehpr2 code
