+++ apps/api/src/unit/unit.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Unit } from 'src/entities/unit.entity';
+import { Unit } from './entity/unit.entity';
 import { UnitController } from './unit.controller';
 import { UnitService } from './unit.service';
